Mini Task 2
Update NavBar;
Add new module Navbar.
Created a method that returns a token from local storag.
Created a method that signs the user out when the signout button is clicked.
Import named function into main view.
Placed navbar in the main view that it appears on all views.
Created the routes for the views in your navbar.